+++ lib/enumerize.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'active_support/concern'
 require 'enumerize/version'
 
@@ -10,6 +12,7 @@ module Enumerize
   autoload :Module,       'enumerize/module'
   autoload :Predicates,   'enumerize/predicates'
   autoload :Predicatable, 'enumerize/predicatable'
+  autoload :Utils,        'enumerize/utils'
   autoload :ModuleAttributes, 'enumerize/module_attributes'
 
   autoload :ActiveModelAttributesSupport, 'enumerize/activemodel'