+++ lib/enumerize/attribute.rb
@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
 module Enumerize
   class Attribute
-    attr_reader :klass, :name, :values, :default_value, :i18n_scope
+    attr_reader :klass, :name, :values, :default_value, :i18n_scope, :skip_validations_value
 
     def initialize(klass, name, options={})
       raise ArgumentError, ':in option is required' unless options[:in]
@@ -26,6 +28,8 @@ module Enumerize
         @default_value = find_default_value(options[:default])
         raise ArgumentError, 'invalid default value' unless @default_value
       end
+
+      @skip_validations_value = options.fetch(:skip_validations, false)
     end
 
     def find_default_value(value)
@@ -42,6 +46,10 @@ module Enumerize
 
     def find_values(*values)
       values.map { |value| find_value(value) }.compact
+    end
+
+    def each_value
+      values.each { |value| yield value }
     end
 
     def i18n_scopes