+++ lib/enumerize/integrations/rspec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'enumerize/integrations/rspec/matcher'
 
 module Enumerize