+++ lib/enumerize/predicatable.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Enumerize
   module Predicatable
     def respond_to_missing?(method, include_private=false)