+++ test/support/mock_controller.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class MockController
   attr_writer :action_name